Code style changes
